simplificação pastas e refatoração buttons
