+++ src/app/layout.tsx
@@ -1,11 +1,16 @@
-import '@/app/ui/global.css';
-import { roboto } from '../app/ui/fonts';
+import { Roboto } from 'next/font/google'
+
+const roboto = Roboto({
+  weight: ['400', '500', '700'],
+  subsets: ['latin'],
+})
 
 export default function RootLayout({
   children,
 }: {
   children: React.ReactNode;
 }) {
+
   return (
     <html lang="en">
       <body className={`${roboto.className} antialiased`}>{children}</body>
