--- src/app/lib/data.ts
@@ -1,1 +1,0 @@
-/* import { sql } from '@vercel/postgres'; */
