--- src/app/ui/dashboard/sidebar.tsx
@@ -1,1 +1,0 @@
-import { FacebookButton, InstagramButton, TikTokButton, SpotifyButton, YouTubeButton } from '../dashboard/mediaButtons';
