--- src/app/ui/fonts.ts
@@ -1,6 +1,0 @@
-import { Roboto } from 'next/font/google';
- 
-export const roboto = Roboto({
-    weight: ['400', '700'],
-    subsets: ['latin'],
-  });
